Add mode support to keybindings
Instead of globally kidnapping the v, h, and t keys, Xest will be smart
enough to online kidnap a key if it's in the right mode. This means that
keybindings used in normal mode won't be grabbed in insert mode.

The current implementation regrabs every key after every event which is
not ideal. It also does not implement temporary modes when the key is
held down.