Massive refactoring and window hiding/showing
First, the project has been refactored heavily. Everything is now event based.
Any new types are stored in Types.hs and most functions are stored in Core.hs.
Splitting up Core.hs would be cool but there is tons of coupling in there that
needs to be fixed.

Events are implemented using Events and Triggers. Triggers are super simple data
types that make use of type families to ensure that a given event always matches
a given trigger. For example, you can't create a KeyEvent using an XorgTrigger.

Windows can now be shown or hidden from the config and changes to the mode can
trigger actions. As a result, there also needed to be a way to specify which
window to modify so Xest can now search for a window with a given class name.
Unfortunately, the Haskell bindings around this are very weak (or I am missing
something) so there are mallocs and other "scary" operations happening. This
should probably be made more safe at some point.

The config file is still an absolute pain to test because of the lack of error
messages. I kind of like the read interface though for the config so if there
was a "readWithErrors" class that could be autoderived, that would be great.