Remove Xest and increase compile times
The previous code was getting hard to manage, and Polysemy looked cool,
so this project now uses more modular effects. Instead of having every
function take in the entire Xest monad, it now takes in Semantics
representing exactly what it needs. Handler should still be broken up
but it's pretty nice. Also changed the handler stuff to have a more
restrictive type signature.

If you change the Base file, get ready for a slightly longer compile
time. It's not unbearable, but is noticable more than the other files.

On the bright side, lens is no longer a dependency because it wasn't
really needed and we use type application now instead.

TODO: Use our newfound powers to make the code testable