Make FocusList more robust
FocusList now keeps track of a visual order and a focus order. This is
extremely similar to how something like i3 works and makes it possible
to define better properties over the Tilers.

Also refactored some code to make Fix vs not fixed more consistent. In
the Polysemy state system, either can be used to access the root Tiler.
This is done because the root Tiler makes sense unfixed, but recursion
schemes expect fixed Tilers.