Add some lenses
This branch holds some experimentation with Lenses and better
documentation. I can't tell if the Lenses make it better or worse to
use this code.

So far, it seems like a lot of functions that were two things have
suddenly become 1. That's kind of cool and powerful, but it comes at the
cost of needing to extract the lenses all over the place.