Totally restructure core and the main loop
Core now only contains helper functions. Everything in handler has been
split up so that we can use Polysemy to more of its potential.

Also, fixed the longstanding bug of the vim/notepad mode not working.
Luckily, you were stuck in vim mode but given that this is one of the
core features of Xest, it's cool that it works.