Improve logging and attempt to fix a race
The logging now looks closer to the pre-refactor version which is nice
for debugging.

There appeared to be a race condition in checkXEvent which was triggered
by closing Termite. When Termite closes, it closes a bunch of random
windows, most of which aren't managed by Xest. For whatever reason, the
window Xest manages is one of the last.

CheckXEvent is used to prevent any rendering from happening while there
are still events that need to be handled. Although we haven't received
the DestroyWindow event for the actual Termite window, the server still
considers it to be dead.

The race condition appears to be caused by the following:

Xest receives one destroy event -> Xest runs checkXEvent -> the event
counter detects a nonempty queue and just returns 1 -> the rest of the
events come in but aren't accounted for.