QOL improvements and new Tilers
First, The config no longer downloads files from the internet. That
ended up being a kind of bad idea for a couple of reasons. Instead, it
expects to find some extra files next to the main config.

Added ZoomMonitorToInput. 99% of the time when I zoom in the monitor, I
want it to zoom all of the way in. This Action does that.

Added ZoomInInputSkip. Most of the time having to step across reverses
and full screens is really annoying. This Action jumps you into the
nearest interesting tiler.

Fixed ZoomOutInput. If the monitor is in the way, both the monitor and
the controller move. Previously, only the monitor would move. The old
behavior was becoming more annoying the more I used it.

Fixed Keybindings. You can now bind the same key in different modes.

Config updates. The config file now has a NormalS mode triggered by the
shift key while in normal mode. The config is roughly organized such
that "smart" actions are done in Normal mode while NormalS has the raw
options. For example, Normal mode # keys zooms you out, changes the
focus, then zooms you back in. This makes simulating workspaces a lot
easier.