WIP: Make motion code less bad
The current motion code is complicated, hard to read, and hard to
modify.

Although it seemed like the complexity was not necessary, I've slowly
been rediscovering why it's all there.

The size parameter is a percentage of the total available space added on
to the default size. This makes drawing and creating horizontal sizes
hard. With that said, it has some positive properties. For example,
sizes are completely independent of the number of windows. This means
you can add or remove windows and the relative sizes stay the same.

Modifying the size of one Tiler in a horizontal requires changing the
sizes of those next to it. This behavior is good, but the FocusList data
type makes this awful. We have to combine the focused order and the
visual order to make it work.

On the bright side, using optics for FocusList seems like a cool idea. I
also really want to get rid of focused list though and replace it with
something better.

Also, I thing I found the bug that triggered this refactoring so I'll
try fixing that instead.