Big refactor and stabality improvements
The biggest change is the new file hierarchy which splits up Base and
removes the Type.hs file. Thanks to Type.hs being gone, Config is now
significantly simpler.

Modes no longer have intro or exit actions. Keeping them would have made
the dhall config significantly more complicated. If you want to perform
some action when entering or exiting a mode, just run the actions on the
key press that changes the modes.

Mapping a window is now more stable. The new window will actually be
drawn instead of ignored for a little while.

Click to focus is back and better than ever. To be fair, that's not hard
since it was previously causing the entire WM to be unusable. The
solution was to find the magic series of parameters and xlib calls that
did the right thing. Thanks to XMonad and a stackoverflow comment for
helping me out there.