Add fullscreen support and fix some EWMH stuff
The big feature here is that Xest now supports fullscreen windows. This
was actually pretty easy. Basically, it just zooms in on whatever wants
to be fullscreen.

The other big feature is that Xest now actually advertises itself as
being standards compliant. The old _NET_SUPPORTING_WM_CHECK code was
completely wrong...