Make Mover more stable
The mover and setFocus functions now try to avoid crashing Xlib.
Theoretically, they shouldn't have been anyway, but things get weird
when you're talking to a shared, stateful, asynchronous api.

It also means that less of the implementation details about Xlib leak
through to the business logic.

I really need more newtype wrappers. I spent an hour or so debugging a
problem with focus when I was using get instead of input. Normally, that
would cause compiler errors, but because all of the X types are just
aliases for Int*, it compiled fine.