Improved logging
It's still far from perfect, but the logging now uses CoLog. I use Colog
in a way that's nearly identical to PrintMe so that's not too great.
What is great is that all State changes are now logged automatically.
Also, it should be easy to improve logging in the future because Colog
looks cool.

There's also some initial work on getting the logger to write messages
on an error, but I can't figure out how to catch async exceptions inside
of Sem.