What Was Sundered and Undone...
At some point in the past, the Horizontal tiler took a parameter which
specified whether it should be tiled along the X, Y, or Z axes. After
seeing some XMonad code, I decided to abandon that and to instead split
X, Y, and Z into distinct Tilers that composed to create cool effects.
Unfortunately, that actually led to a lot of weird edge cases and was
a lot less elegant than I wanted.

Now, I've recombined those 3 Tilers and managed to add Floating in as
well. A single "Many" Tiler can be configured for all of the 3 use cases
for Horiz and for Floating as well. Not only does this fix some bugs,
but it also significantly reduces code duplication and makes Floating a
lot less sketchy. This seems like it will be a major win for the
project.