Move borders and screen size into tree
Borders and screen sizes are now stored by the Tilers that use them.
This was originally part of adding struts to Xest. Unfortunately, struts
are kind of hard but this was a nice refactor anyway. This is also kind
of WIP so maybe don't use this commit for anything real.