Remove ClassyPrelude and NonEmpty
ClassyPrelude is gone which means we no longer have the Mono hierarchy
to deal with. Nonempty has also been deleted and replaced with the
version in Base. I don't remember why I created NonEmpty a while ago,
but it turns out we didn't need it.

I've also started the slow lensification of Xest. This time I think it
will actually be a good idea unlike the last several times when I've
tried.