Improve rendering in the face of asynchronicity
Imagine you have a lot of narrow windows and you quickly move your mouse
across all of them. Xest will receive the following sequence of events:
X-X-X-X-X-X where every X is an EnterNotify event.

When Xest gets an EnterNotify event, it changes the focused window and
calls the refresh function. That looks like the following (scale is
roughly the same as above)
X--YX--YX--Y Where X is the start of rendering and Y is the end

If you line up the two timelines, there's a problem. While rendering,
Xest ignores all EnterNotify events. We need to do this because things
like changing workspaces will trigger EnterNotify events which messes
with the focus.

Since X11 is asynchronous, there's a pretty good chance of ignoring
legitimate EnterNotify events as well. The solution for now is to define
Safe and Unsafe redraws. a safe redraw is one which will never create
additional EnterNotify events. The only event that meets this
requirement is an EnterNotify event. All other events might change the
size or composition of the windows on the screen. Xest only ignores
EnterNotify events during an UnsafeRedraw.

While this still leaves race conditions in the code it fixes the biggest
offender.

I'm not sure why it sometimes still misses intermediate windows, but it
catches the last one so I guess that's fine. It's also really slow for
some reason.