Remove the pop function from focusedList
Why remove the pop function? At a high level, FocusedList is a list of
elements with multiple orderings. When you describe it like that, it
sounds really simple. The current export list leaks all kinds of nasty
details and the pop function was one of them. TBH I don't remember half
of what's going on in that module which makes maintaining it a scary
proposition right now. Using lenses will hopefully fix that.