Switch to freer-simple
Capabilities was cool, but it doesn't build on GHC 8.10. Since
performance isn't that important, I'm hoping free-simple will be easier
to use and have more support.

Luckily, the refactor was pretty painless and actually ran correctly as
soon as it compiled. There might be some extra stuff in DoAll but I'll
remove those as needed.