Try to capture configure events
A lot of dialog windows are setting really bad default sizes. My current
guess is that they're trying to use configureWindow to pick a more
reasonable size but I'm completely ignoring that call. I need to figure
out how to let apps configure their windows when the window is floating.

Floating here means an immediate child of a Floating Tiler.